+++ source/Energinet.DataHub.MeteringPoints.Application/UpdateMasterData/UpdateMasterDataHandler.cs
@@ -20,6 +20,7 @@ using System.Threading.Tasks;
 using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Application.Common.ChildMeteringPoints;
 using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
+using Energinet.DataHub.MeteringPoints.Domain.BusinessProcesses;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.Policies;
@@ -34,8 +35,7 @@ namespace Energinet.DataHub.MeteringPoints.Application.ChangeMasterData
         private readonly ChangeMasterDataSettings _settings;
         private readonly ChangeMeteringPointAuthorizer _authorizer;
         private readonly ParentChildCouplingHandler _parentChildCouplingHandler;
-        private readonly MasterDataValidator _masterDataValidator;
-        private readonly MasterDataUpdateHandler _masterDataUpdateHandler;
+        private readonly UpdateMasterDataProcess _updateMasterDataProcess;
 
         public ChangeMasterDataHandler(
             IMeteringPointRepository meteringPointRepository,
@@ -43,16 +43,14 @@ namespace Energinet.DataHub.MeteringPoints.Application.ChangeMasterData
             ChangeMasterDataSettings settings,
             ChangeMeteringPointAuthorizer authorizer,
             ParentChildCouplingHandler parentChildCouplingHandler,
-            MasterDataValidator masterDataValidator,
-            MasterDataUpdateHandler masterDataUpdateHandler)
+            UpdateMasterDataProcess updateMasterDataProcess)
         {
             _meteringPointRepository = meteringPointRepository ?? throw new ArgumentNullException(nameof(meteringPointRepository));
             _systemDateTimeProvider = systemDateTimeProvider ?? throw new ArgumentNullException(nameof(systemDateTimeProvider));
             _settings = settings;
             _authorizer = authorizer ?? throw new ArgumentNullException(nameof(authorizer));
             _parentChildCouplingHandler = parentChildCouplingHandler;
-            _masterDataValidator = masterDataValidator ?? throw new ArgumentNullException(nameof(masterDataValidator));
-            _masterDataUpdateHandler = masterDataUpdateHandler ?? throw new ArgumentNullException(nameof(masterDataUpdateHandler));
+            _updateMasterDataProcess = updateMasterDataProcess ?? throw new ArgumentNullException(nameof(updateMasterDataProcess));
         }
 
         public async Task<BusinessProcessResult> Handle(ChangeMasterDataRequest request, CancellationToken cancellationToken)
@@ -90,13 +88,13 @@ namespace Energinet.DataHub.MeteringPoints.Application.ChangeMasterData
                 return new BusinessProcessResult(request.TransactionId, policyCheckResult.Errors);
             }
 
-            var validationResult = await _masterDataUpdateHandler.CanUpdateAsync(targetMeteringPoint, updatedMasterData).ConfigureAwait(false);
+            var validationResult = await _updateMasterDataProcess.CanUpdateAsync(targetMeteringPoint, updatedMasterData).ConfigureAwait(false);
             if (validationResult.Success != true)
             {
                 return new BusinessProcessResult(request.TransactionId, validationResult.Errors);
             }
 
-            _masterDataUpdateHandler.Update(targetMeteringPoint, updatedMasterData);
+            _updateMasterDataProcess.Update(targetMeteringPoint, updatedMasterData);
 
             var parentCouplingResult = await HandleParentChildCouplingAsync(request, targetMeteringPoint).ConfigureAwait(false);
             return parentCouplingResult.Success == false ? parentCouplingResult : BusinessProcessResult.Ok(request.TransactionId);
